Update PawnKindDefs to A16
I diffed the PawnKind classes with the core definitions and merged the
changes necessary for this mod.

Balance-relevant attributes such as weaponMoney and
baseRecruitDifficulty were left alone in the few cases they conflicted
with the values in core.

Base classes were abstracted out in places to correspond to those in
core.

The principal changes are that overally, pawns carry more food, may
carry silver, drugs, or medicine, and have a chance to be addicted.

TribalArcherPila was removed from PawnKinds_Tribal.xml as it was not
present in core and isn't referenced in the Tribal faction.

The most significant divergence from core is that core has mercenary
grenadiers, which are not present in this mod.

Overall, these definitions now diff cleanly with core and it should be
easy to update them in future as balance changes are made to the base game.