Make ranged and melee weapons smeltable
Ranged and melee weapons are now smeltable. The amount of resources
returned appears to be the production cost of the weapon, modified by
and efficency factor of 0.25. I'm not sure where the efficiency factor
is set.

Turrets, grenades and ammunition are not smeltable. The production of
these is defined through a RecipeDef external to the ThingDef, whereas
the production of guns and melee weapons is defined through a
recipeMaker and costList internal to the ThingDef. This does not
appear to work with the method of setting the Thing to smeltable=true.

Smelting grenades and ammunition has fundamental problems as the game
does not appear to provide a way of smelting stacks, and those are
produced in stacks. Besides not having the smelting products
automatically determined by the production cost, having to smelt each
item in the stack individually seriously unbalances the work cost and
makes balancing the resources difficult.

Making turrets smeltable with the proper amount of resources returned
is likely possible. However, having them no longer smeltable is not a
regression as smelting turrets did not previously return any
resources, anyway.

In short, the only things smeltable are ranged and melee weapons, but
everything smeltable should return the correct amount and type of
resources.

Fixes #45