Make CarryBulk and CarryWeight a function of pawn BodySize
The defaultBaseValues for these stats have been changed to 0. Setting
the stat on a pawn definition acts as an offset.

This change is principally for consistency with
RimWorld.MassUtility.Capacity, in which mass capacity is produced as a
linear function of pawn body size.

The proportionality constant for CarryWeight is the same as in
MassUtility, and the offsets and proportionality constant for
CarryBulk were chosen to have minimal impact on balance.

A consequence is that as the proportionality constant for
CarryWeight (MassUtility.MassCapacityPerBodySize) is 35, and a human
has a baseBodySize of 1, CarryWeight for a baseline human has been
reduced from 40 to 35.